move field validation functions to new file and make post api requsts work
