+++ src/alliance/SignupForm.tsx
@@ -3,7 +3,8 @@ import { StyleSheet, View } from 'react-native'
 import { NewMember } from 'src/alliance/AllianceMember'
 import { CheckboxWithLabel } from 'src/forms/CheckboxWithLabel'
 import { ErrorDisplay } from 'src/forms/ErrorDisplay'
-import FormContainer, { emailIsValid, hasField } from 'src/forms/Form'
+import FormContainer from 'src/forms/Form'
+import { emailIsValid, hasField } from "src/forms/emailIsValid"
 import { Form } from 'src/forms/FormComponents'
 import { LabeledInput } from 'src/forms/LabeledInput'
 import SubmitButton from 'src/forms/SubmitButton'
