+++ src/community/EcoFund.tsx
@@ -4,7 +4,8 @@ import ah from 'src/community/ah-logo.png'
 import polychain from 'src/community/polychain-logo.png'
 import { H2 } from 'src/fonts/Fonts'
 import { ErrorDisplay, ErrorKeys } from 'src/forms/ErrorDisplay'
-import FormContainer, { emailIsValid, hasField } from 'src/forms/Form'
+import FormContainer from 'src/forms/Form'
+import { emailIsValid, hasField } from "src/forms/emailIsValid"
 import { Form } from 'src/forms/FormComponents'
 import { LabeledInput } from 'src/forms/LabeledInput'
 import SubmitButton from 'src/forms/SubmitButton'
