+++ src/community/connect/FellowshipForm.tsx
@@ -2,7 +2,8 @@ import { FellowAppShape } from 'fullstack/Fellowship'
 import * as React from 'react'
 import { StyleSheet } from 'react-native'
 import { ErrorDisplay } from 'src/forms/ErrorDisplay'
-import FormContainer, { emailIsValid, hasField } from 'src/forms/Form'
+import FormContainer from 'src/forms/Form'
+import { emailIsValid, hasField } from "src/forms/emailIsValid"
 import { Form } from 'src/forms/FormComponents'
 import { LabeledInput } from 'src/forms/LabeledInput'
 import SubmitButton from 'src/forms/SubmitButton'
